Refactor LoginLocalStorage
